+++ src/advanced/__tests__/advanced.test.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
 import { CartPage } from '../../refactoring/components/cart/CartPage';
-import { AdminPage } from '../../refactoring/components/AdminPage';
+import { AdminPage } from '../../refactoring/components/admin/AdminPage';
 import { Coupon, Product } from '../../types';
 
 const mockProducts: Product[] = [
